The install recipe now calls other recipes directly instead of calling default
The sentinel recipe had been causing the configure recipe to fire. This was a
problem as it would configure a default redis instance when you probably didn't
want it to.  You would end up with sentinel installed and running but config
cruft left over that shouldn't have been there.  This solves that problem!