Fixes swapfile name bug and fixes logfile directory and file bugs
If the directory for the logfile wasn't created, the cookbook was exploding.
While fixing that I realized that redis 2.8 also changed how it handles the
configuration for stdout.  It now uses the empty string.

The swapfile wasn't being named with the server name, so if you had 2 servers
running in parallel they would have shared a swapfile.  This only effects redis
2.4 or lower.  Now each instance has it's own swapfile.