Adds required-start/required-stop functionality, Default redis install is now 2.8, Fixes bug in log_file setup
required-start and required-stop default to empty strings.  This does allow a
user to specify required stop and starts for their redis instances when using
init.d scripts.  Accepts an array of strings

Changes the default version of redis to 2.8.1

Warns a user if they setup a logfile but also log to the system logger.