Changes job_control to be a top level attribute.  Fixes status issues with upstart and init.d
The job_control attribute is now a top level one and applies to all instances of
redis and sentinel that are running.

When using upstart, the status command now relies on the default status method
(which uses /sbin/status)

When using init.d, I have added a status command to the init script and the init.d
based service now uses that instead of trying to use pgrep.

Minor cleanup to the README file and additions to the changelog.

Bumps redis default version to 2.8.2