keyspace version conditional, actually use param
Introduced in 0d4553be4221025f794e96951ace53baefa64f74 and
brianbianco/redisio#106

This configuration item was introduced in the 2.8.x series of Redis, and
using the cookbook on older redis nodes breaks the configuration, as the
config is written and service startup is halted due to unrecognized
entry.

This also adds support for actually using the parameter passed, as
previously it remianed an empty string no matter what you would pass to
the attribute.